requirements.txt
requests
pyyaml
packageurl-python
semantic_version
colorama